+++ pycvvdp/video_source_file.py
@@ -3,7 +3,7 @@
 from asyncio.log import logger
 from functools import cache
 import os
-from turtle import color
+# from turtle import color
 import imageio.v2 as io
 import numpy as np
 from torch.functional import Tensor